Comments were fixed fro Vlad
